cbor gen
