+++ builtin/v12/miner/cbor_gen.go
@@ -11,7 +11,6 @@ import (
 	address "github.com/filecoin-project/go-address"
 	abi "github.com/filecoin-project/go-state-types/abi"
 	verifreg "github.com/filecoin-project/go-state-types/builtin/v12/verifreg"
-	exitcode "github.com/filecoin-project/go-state-types/exitcode"
 	proof "github.com/filecoin-project/go-state-types/proof"
 	cid "github.com/ipfs/go-cid"
 	cbg "github.com/whyrusleeping/cbor-gen"
@@ -6389,1088 +6388,6 @@ func (t *GetMultiAddrsReturn) UnmarshalCBOR(r io.Reader) (err error) {
 	}
 
 	if _, err := io.ReadFull(cr, t.MultiAddrs[:]); err != nil {
-		return err
-	}
-	return nil
-}
-
-var lengthBufProveCommitSectors2Params = []byte{133}
-
-func (t *ProveCommitSectors2Params) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufProveCommitSectors2Params); err != nil {
-		return err
-	}
-
-	// t.SectorActivations ([]miner.SectorActivationManifest) (slice)
-	if len(t.SectorActivations) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.SectorActivations was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.SectorActivations))); err != nil {
-		return err
-	}
-	for _, v := range t.SectorActivations {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-
-	// t.SectorProofs ([][]uint8) (slice)
-	if len(t.SectorProofs) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.SectorProofs was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.SectorProofs))); err != nil {
-		return err
-	}
-	for _, v := range t.SectorProofs {
-		if len(v) > cbg.ByteArrayMaxLen {
-			return xerrors.Errorf("Byte array in field v was too long")
-		}
-
-		if err := cw.WriteMajorTypeHeader(cbg.MajByteString, uint64(len(v))); err != nil {
-			return err
-		}
-
-		if _, err := cw.Write(v[:]); err != nil {
-			return err
-		}
-	}
-
-	// t.AggregateProof ([]uint8) (slice)
-	if len(t.AggregateProof) > cbg.ByteArrayMaxLen {
-		return xerrors.Errorf("Byte array in field t.AggregateProof was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajByteString, uint64(len(t.AggregateProof))); err != nil {
-		return err
-	}
-
-	if _, err := cw.Write(t.AggregateProof[:]); err != nil {
-		return err
-	}
-
-	// t.RequireActivationSuccess (bool) (bool)
-	if err := cbg.WriteBool(w, t.RequireActivationSuccess); err != nil {
-		return err
-	}
-
-	// t.RequireNotificationSuccess (bool) (bool)
-	if err := cbg.WriteBool(w, t.RequireNotificationSuccess); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *ProveCommitSectors2Params) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = ProveCommitSectors2Params{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 5 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.SectorActivations ([]miner.SectorActivationManifest) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.SectorActivations: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.SectorActivations = make([]SectorActivationManifest, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v SectorActivationManifest
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.SectorActivations[i] = v
-	}
-
-	// t.SectorProofs ([][]uint8) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.SectorProofs: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.SectorProofs = make([][]uint8, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-		{
-			var maj byte
-			var extra uint64
-			var err error
-
-			maj, extra, err = cr.ReadHeader()
-			if err != nil {
-				return err
-			}
-
-			if extra > cbg.ByteArrayMaxLen {
-				return fmt.Errorf("t.SectorProofs[i]: byte array too large (%d)", extra)
-			}
-			if maj != cbg.MajByteString {
-				return fmt.Errorf("expected byte array")
-			}
-
-			if extra > 0 {
-				t.SectorProofs[i] = make([]uint8, extra)
-			}
-
-			if _, err := io.ReadFull(cr, t.SectorProofs[i][:]); err != nil {
-				return err
-			}
-		}
-	}
-
-	// t.AggregateProof ([]uint8) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.ByteArrayMaxLen {
-		return fmt.Errorf("t.AggregateProof: byte array too large (%d)", extra)
-	}
-	if maj != cbg.MajByteString {
-		return fmt.Errorf("expected byte array")
-	}
-
-	if extra > 0 {
-		t.AggregateProof = make([]uint8, extra)
-	}
-
-	if _, err := io.ReadFull(cr, t.AggregateProof[:]); err != nil {
-		return err
-	}
-	// t.RequireActivationSuccess (bool) (bool)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajOther {
-		return fmt.Errorf("booleans must be major type 7")
-	}
-	switch extra {
-	case 20:
-		t.RequireActivationSuccess = false
-	case 21:
-		t.RequireActivationSuccess = true
-	default:
-		return fmt.Errorf("booleans are either major type 7, value 20 or 21 (got %d)", extra)
-	}
-	// t.RequireNotificationSuccess (bool) (bool)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajOther {
-		return fmt.Errorf("booleans must be major type 7")
-	}
-	switch extra {
-	case 20:
-		t.RequireNotificationSuccess = false
-	case 21:
-		t.RequireNotificationSuccess = true
-	default:
-		return fmt.Errorf("booleans are either major type 7, value 20 or 21 (got %d)", extra)
-	}
-	return nil
-}
-
-var lengthBufSectorActivationManifest = []byte{130}
-
-func (t *SectorActivationManifest) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufSectorActivationManifest); err != nil {
-		return err
-	}
-
-	// t.SectorNumber (abi.SectorNumber) (uint64)
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajUnsignedInt, uint64(t.SectorNumber)); err != nil {
-		return err
-	}
-
-	// t.Pieces ([]miner.PieceActivationManifest) (slice)
-	if len(t.Pieces) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.Pieces was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.Pieces))); err != nil {
-		return err
-	}
-	for _, v := range t.Pieces {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *SectorActivationManifest) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = SectorActivationManifest{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.SectorNumber (abi.SectorNumber) (uint64)
-
-	{
-
-		maj, extra, err = cr.ReadHeader()
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.SectorNumber = abi.SectorNumber(extra)
-
-	}
-	// t.Pieces ([]miner.PieceActivationManifest) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.Pieces: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.Pieces = make([]PieceActivationManifest, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v PieceActivationManifest
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.Pieces[i] = v
-	}
-
-	return nil
-}
-
-var lengthBufPieceActivationManifest = []byte{132}
-
-func (t *PieceActivationManifest) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufPieceActivationManifest); err != nil {
-		return err
-	}
-
-	// t.CID (cid.Cid) (struct)
-
-	if err := cbg.WriteCid(cw, t.CID); err != nil {
-		return xerrors.Errorf("failed to write cid field t.CID: %w", err)
-	}
-
-	// t.Size (abi.PaddedPieceSize) (uint64)
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajUnsignedInt, uint64(t.Size)); err != nil {
-		return err
-	}
-
-	// t.VerifiedAllocationKey (miner.VerifiedAllocationKey) (struct)
-	if err := t.VerifiedAllocationKey.MarshalCBOR(cw); err != nil {
-		return err
-	}
-
-	// t.Notify ([]miner.DataActivationNotification) (slice)
-	if len(t.Notify) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.Notify was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.Notify))); err != nil {
-		return err
-	}
-	for _, v := range t.Notify {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *PieceActivationManifest) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = PieceActivationManifest{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 4 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.CID (cid.Cid) (struct)
-
-	{
-
-		c, err := cbg.ReadCid(cr)
-		if err != nil {
-			return xerrors.Errorf("failed to read cid field t.CID: %w", err)
-		}
-
-		t.CID = c
-
-	}
-	// t.Size (abi.PaddedPieceSize) (uint64)
-
-	{
-
-		maj, extra, err = cr.ReadHeader()
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.Size = abi.PaddedPieceSize(extra)
-
-	}
-	// t.VerifiedAllocationKey (miner.VerifiedAllocationKey) (struct)
-
-	{
-
-		b, err := cr.ReadByte()
-		if err != nil {
-			return err
-		}
-		if b != cbg.CborNull[0] {
-			if err := cr.UnreadByte(); err != nil {
-				return err
-			}
-			t.VerifiedAllocationKey = new(VerifiedAllocationKey)
-			if err := t.VerifiedAllocationKey.UnmarshalCBOR(cr); err != nil {
-				return xerrors.Errorf("unmarshaling t.VerifiedAllocationKey pointer: %w", err)
-			}
-		}
-
-	}
-	// t.Notify ([]miner.DataActivationNotification) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.Notify: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.Notify = make([]DataActivationNotification, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v DataActivationNotification
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.Notify[i] = v
-	}
-
-	return nil
-}
-
-var lengthBufVerifiedAllocationKey = []byte{130}
-
-func (t *VerifiedAllocationKey) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufVerifiedAllocationKey); err != nil {
-		return err
-	}
-
-	// t.Client (abi.ActorID) (uint64)
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajUnsignedInt, uint64(t.Client)); err != nil {
-		return err
-	}
-
-	// t.ID (verifreg.AllocationId) (uint64)
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajUnsignedInt, uint64(t.ID)); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (t *VerifiedAllocationKey) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = VerifiedAllocationKey{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Client (abi.ActorID) (uint64)
-
-	{
-
-		maj, extra, err = cr.ReadHeader()
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.Client = abi.ActorID(extra)
-
-	}
-	// t.ID (verifreg.AllocationId) (uint64)
-
-	{
-
-		maj, extra, err = cr.ReadHeader()
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.ID = verifreg.AllocationId(extra)
-
-	}
-	return nil
-}
-
-var lengthBufDataActivationNotification = []byte{130}
-
-func (t *DataActivationNotification) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufDataActivationNotification); err != nil {
-		return err
-	}
-
-	// t.Address (address.Address) (struct)
-	if err := t.Address.MarshalCBOR(cw); err != nil {
-		return err
-	}
-
-	// t.Payload ([]uint8) (slice)
-	if len(t.Payload) > cbg.ByteArrayMaxLen {
-		return xerrors.Errorf("Byte array in field t.Payload was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajByteString, uint64(len(t.Payload))); err != nil {
-		return err
-	}
-
-	if _, err := cw.Write(t.Payload[:]); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *DataActivationNotification) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = DataActivationNotification{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Address (address.Address) (struct)
-
-	{
-
-		if err := t.Address.UnmarshalCBOR(cr); err != nil {
-			return xerrors.Errorf("unmarshaling t.Address: %w", err)
-		}
-
-	}
-	// t.Payload ([]uint8) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.ByteArrayMaxLen {
-		return fmt.Errorf("t.Payload: byte array too large (%d)", extra)
-	}
-	if maj != cbg.MajByteString {
-		return fmt.Errorf("expected byte array")
-	}
-
-	if extra > 0 {
-		t.Payload = make([]uint8, extra)
-	}
-
-	if _, err := io.ReadFull(cr, t.Payload[:]); err != nil {
-		return err
-	}
-	return nil
-}
-
-var lengthBufProveCommit2Return = []byte{129}
-
-func (t *ProveCommit2Return) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufProveCommit2Return); err != nil {
-		return err
-	}
-
-	// t.Sectors ([]miner.SectorActivationReturn) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.Sectors was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.Sectors))); err != nil {
-		return err
-	}
-	for _, v := range t.Sectors {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ProveCommit2Return) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = ProveCommit2Return{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 1 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Sectors ([]miner.SectorActivationReturn) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.Sectors: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.Sectors = make([]SectorActivationReturn, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v SectorActivationReturn
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.Sectors[i] = v
-	}
-
-	return nil
-}
-
-var lengthBufSectorActivationReturn = []byte{131}
-
-func (t *SectorActivationReturn) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufSectorActivationReturn); err != nil {
-		return err
-	}
-
-	// t.Activated (bool) (bool)
-	if err := cbg.WriteBool(w, t.Activated); err != nil {
-		return err
-	}
-
-	// t.Power (big.Int) (struct)
-	if err := t.Power.MarshalCBOR(cw); err != nil {
-		return err
-	}
-
-	// t.Pieces ([]miner.PieceActivationReturn) (slice)
-	if len(t.Pieces) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.Pieces was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.Pieces))); err != nil {
-		return err
-	}
-	for _, v := range t.Pieces {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *SectorActivationReturn) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = SectorActivationReturn{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 3 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Activated (bool) (bool)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajOther {
-		return fmt.Errorf("booleans must be major type 7")
-	}
-	switch extra {
-	case 20:
-		t.Activated = false
-	case 21:
-		t.Activated = true
-	default:
-		return fmt.Errorf("booleans are either major type 7, value 20 or 21 (got %d)", extra)
-	}
-	// t.Power (big.Int) (struct)
-
-	{
-
-		if err := t.Power.UnmarshalCBOR(cr); err != nil {
-			return xerrors.Errorf("unmarshaling t.Power: %w", err)
-		}
-
-	}
-	// t.Pieces ([]miner.PieceActivationReturn) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.Pieces: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.Pieces = make([]PieceActivationReturn, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v PieceActivationReturn
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.Pieces[i] = v
-	}
-
-	return nil
-}
-
-var lengthBufPieceActivationReturn = []byte{130}
-
-func (t *PieceActivationReturn) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufPieceActivationReturn); err != nil {
-		return err
-	}
-
-	// t.Claimed (bool) (bool)
-	if err := cbg.WriteBool(w, t.Claimed); err != nil {
-		return err
-	}
-
-	// t.Notifications ([]miner.DataActivationNotificationReturn) (slice)
-	if len(t.Notifications) > cbg.MaxLength {
-		return xerrors.Errorf("Slice value in field t.Notifications was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajArray, uint64(len(t.Notifications))); err != nil {
-		return err
-	}
-	for _, v := range t.Notifications {
-		if err := v.MarshalCBOR(cw); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *PieceActivationReturn) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = PieceActivationReturn{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Claimed (bool) (bool)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajOther {
-		return fmt.Errorf("booleans must be major type 7")
-	}
-	switch extra {
-	case 20:
-		t.Claimed = false
-	case 21:
-		t.Claimed = true
-	default:
-		return fmt.Errorf("booleans are either major type 7, value 20 or 21 (got %d)", extra)
-	}
-	// t.Notifications ([]miner.DataActivationNotificationReturn) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.MaxLength {
-		return fmt.Errorf("t.Notifications: array too large (%d)", extra)
-	}
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("expected cbor array")
-	}
-
-	if extra > 0 {
-		t.Notifications = make([]DataActivationNotificationReturn, extra)
-	}
-
-	for i := 0; i < int(extra); i++ {
-
-		var v DataActivationNotificationReturn
-		if err := v.UnmarshalCBOR(cr); err != nil {
-			return err
-		}
-
-		t.Notifications[i] = v
-	}
-
-	return nil
-}
-
-var lengthBufDataActivationNotificationReturn = []byte{130}
-
-func (t *DataActivationNotificationReturn) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-
-	cw := cbg.NewCborWriter(w)
-
-	if _, err := cw.Write(lengthBufDataActivationNotificationReturn); err != nil {
-		return err
-	}
-
-	// t.Code (exitcode.ExitCode) (int64)
-	if t.Code >= 0 {
-		if err := cw.WriteMajorTypeHeader(cbg.MajUnsignedInt, uint64(t.Code)); err != nil {
-			return err
-		}
-	} else {
-		if err := cw.WriteMajorTypeHeader(cbg.MajNegativeInt, uint64(-t.Code-1)); err != nil {
-			return err
-		}
-	}
-
-	// t.Data ([]uint8) (slice)
-	if len(t.Data) > cbg.ByteArrayMaxLen {
-		return xerrors.Errorf("Byte array in field t.Data was too long")
-	}
-
-	if err := cw.WriteMajorTypeHeader(cbg.MajByteString, uint64(len(t.Data))); err != nil {
-		return err
-	}
-
-	if _, err := cw.Write(t.Data[:]); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *DataActivationNotificationReturn) UnmarshalCBOR(r io.Reader) (err error) {
-	*t = DataActivationNotificationReturn{}
-
-	cr := cbg.NewCborReader(r)
-
-	maj, extra, err := cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-	}()
-
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Code (exitcode.ExitCode) (int64)
-	{
-		maj, extra, err := cr.ReadHeader()
-		var extraI int64
-		if err != nil {
-			return err
-		}
-		switch maj {
-		case cbg.MajUnsignedInt:
-			extraI = int64(extra)
-			if extraI < 0 {
-				return fmt.Errorf("int64 positive overflow")
-			}
-		case cbg.MajNegativeInt:
-			extraI = int64(extra)
-			if extraI < 0 {
-				return fmt.Errorf("int64 negative oveflow")
-			}
-			extraI = -1 - extraI
-		default:
-			return fmt.Errorf("wrong type for int64 field: %d", maj)
-		}
-
-		t.Code = exitcode.ExitCode(extraI)
-	}
-	// t.Data ([]uint8) (slice)
-
-	maj, extra, err = cr.ReadHeader()
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.ByteArrayMaxLen {
-		return fmt.Errorf("t.Data: byte array too large (%d)", extra)
-	}
-	if maj != cbg.MajByteString {
-		return fmt.Errorf("expected byte array")
-	}
-
-	if extra > 0 {
-		t.Data = make([]uint8, extra)
-	}
-
-	if _, err := io.ReadFull(cr, t.Data[:]); err != nil {
 		return err
 	}
 	return nil
